+++ src/main/java/dao/dish/IDishStatDAO.java
@@ -6,7 +6,7 @@ import java.util.ArrayList;
 public interface IDishStatDAO {
     ArrayList<DishStat> getBestSellingDish(String startDate, String endDate);
 
-    double getTotalRevenue(String startDate, String endDate);
+    int getTotalRevenue(String startDate, String endDate);
 
     ArrayList<DishStat> getTopSellingDish(String startDate, String endDate, int topN);
 }
