+++ src/main/java/model/MonthlyRevenueStat.java
@@ -5,13 +5,13 @@ import java.io.Serializable;
 public class MonthlyRevenueStat implements Serializable {
     private int month;
     private int year;
-    private double totalRevenue;
+    private int totalRevenue;
 
     public MonthlyRevenueStat() {
         super();
     }
 
-    public MonthlyRevenueStat(int month, int year, double totalRevenue) {
+    public MonthlyRevenueStat(int month, int year, int totalRevenue) {
         this.month = month;
         this.year = year;
         this.totalRevenue = totalRevenue;
@@ -23,6 +23,6 @@ public class MonthlyRevenueStat implements Serializable {
     public int getYear() { return year; }
     public void setYear(int year) { this.year = year; }
 
-    public double getTotalRevenue() { return totalRevenue; }
-    public void setTotalRevenue(double totalRevenue) { this.totalRevenue = totalRevenue; }
+    public int getTotalRevenue() { return totalRevenue; }
+    public void setTotalRevenue(int totalRevenue) { this.totalRevenue = totalRevenue; }
 }
